rename types/resouces.ts to account.ts
