+++ app/javascript/mastodon/components/avatar_overlay.tsx
@@ -1,5 +1,5 @@
 import { useHovering } from '../../hooks/useHovering';
-import type { Account } from '../../types/resources';
+import type { Account } from '../../types/account';
 import { autoPlayGif } from '../initial_state';
 
 interface Props {
