+++ app/javascript/mastodon/components/display_name.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import type { List } from 'immutable';
 
-import type { Account } from '../../types/resources';
+import type { Account } from '../../types/account';
 import { autoPlayGif } from '../initial_state';
 
 import { Skeleton } from './skeleton';
